CKeditor/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-taggit==2.0.0
sqlparse==0.4.2
tzdata==2021.5
